demo/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
transformers>=4.15.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.0.0

# API dependencies
flask>=2.0.0
flask-cors>=3.0.0

# Optional: GPU support (uncomment if using GPU)
# faiss-gpu>=1.7.0

# Development dependencies (optional)
# pytest>=6.0.0
# requests>=2.25.0
